% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomous.R
\name{dichotomous}
\alias{dichotomous}
\title{Function dichotomous}
\usage{
dichotomous(nrow = 1000, probs = c(0.5, 0.6, 0.2, 0.3, 0.4), cor12 = 0.25)
}
\arguments{
\item{nrow}{number of rows in the output dataset}

\item{probs}{a numeric vector of 5 numbers between 0 and 1.
These are the approximate column means of the output dataset.}

\item{cor12}{correlation between x1 and x2. Must be at least 0
and less than 1.}
}
\description{
Simulate a dataset with
5 interdependent dichotomous variables.
}
\details{
The output dataset has variables/columns x1 through x5.
All variables are dichotomous (values are either 0 or 1).
The marginal means are \code{probs} for x1 through x5,
respectively.
x1 and x2 are correlated with correlation \code{cor12}.
x3 is nested in x1 in the sense that
  1. P(x3 = 1 | x1 = 0) = 0, and
  2. P(x3 = 1 | x1 = 1) = P(x3 = 1 and x1 = 1) / P(x1 = 1)
                        = P(x3 = 1) / P(x1 = 1)
Likewise, x4 is nested in x1 and x5 is nested in x2.
}
