% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tboot-package}
\alias{tboot-package}
\alias{drake}
\title{Twisted Bootstrap.
Weighted resampling of data rows
to achieve target column means.}
\description{
Twisted Bootstrap.
Weighted resampling of data rows
to achieve target column means.
}
\author{
William Michael Landau \email{will.landau@lilly.com}

Nathan Morris \email{morris_nathan@lilly.com}
}
