% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomous.R
\name{dichotomous}
\alias{dichotomous}
\title{Function dichotomous}
\usage{
dichotomous(nrow = 1000, probs = c(0.5, 0.6, 0.2, 0.3, 0.4),
  prob12 = 0.4)
}
\arguments{
\item{nrow}{number of rows in the output dataset}

\item{probs}{a numeric vector of 5 numbers between 0 and 1.
These are the marginal probabilities of x1 through x5, respectively
(approximate column means of the output dataset).}

\item{prob12}{Prob(x1 = 1 and x2 = 1)}
}
\description{
Simulate a dataset with
5 interdependent dichotomous variables.
}
\details{
The output dataset has variables/columns x1 through x5.
All variables are dichotomous (values are either 0 or 1).
The marginal means are \code{probs} for x1 through x5,
respectively.
x1 and x2 are correlated such that
Prob(x1 = 1 and x2 = 1) is \code{prob12}.
x3 is nested in x1 in the sense that
  1. Prob(x3 = 1 | x1 = 0) = 0, and
  2. Prob(x3 = 1 | x1 = 1) = Prob(x3 = 1 and x1 = 1) / Prob(x1 = 1)
                           = Prob(x3 = 1) / Prob(x1 = 1)
Likewise, x4 is nested in x1 and x5 is nested in x2.
}
