% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tboot-package}
\alias{tboot-package}
\title{tboot: tilted bootstrapping and Bayesian marginal reconstruction.}
\description{
tboot: tilted bootstrapping and Bayesian marginal reconstruction.
}
\references{
\url{https://github.com/wlandau-lilly/tboot}
}
\author{
Nathan Morris \email{morris_nathan@lilly.com}
}
