% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tboot.R
\name{tboot}
\alias{tboot}
\title{Function tboot}
\usage{
tboot(dataset, weights = rep(1/nrow(dataset), nrow(dataset)),
  nrow = base::nrow(dataset))
}
\arguments{
\item{dataset}{Data frame or matrix to bootstrap.}

\item{weights}{numeric vector of row weights for resampling.
Must have length \code{nrow(dataset)}}

\item{nrow}{number of rows in the new bootstrapped dataset.}
}
\description{
Bootstrap \code{n} rows of \code{dataset} using
weights \code{p}.
}
\seealso{
\code{\link{tweights}}
}
