% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tboot_bmr.R
\name{tboot_bmr}
\alias{tboot_bmr}
\title{Function tboot}
\usage{
tboot_bmr(nrow, bmr, tol_rel_sd = 0.1)
}
\arguments{
\item{nrow}{number of rows in the new bootstrapped dataset.}

\item{dataset}{Data frame or matrix to bootstrap. Rows of the dataset must be in the 
same order as was used for the 'tweights' call. However the dataset may include
additional columns not included in the 'tweights' calll.}

\item{weights}{an object of class 'tweights' output from the 'tweights' function.}

\item{fillMissingAug}{fill in missing augmentation with primary weights resampling.}
}
\description{
Bootstrap \code{nrow} rows of \code{dataset} using
the given row-level weights.
}
\seealso{
\code{\link{tweights}}
}
