% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweights.R
\name{tweights}
\alias{tweights}
\title{Function \code{tweights}}
\usage{
tweights(dataset, target = apply(dataset, 2, mean), max_iter = 1e+05)
}
\arguments{
\item{dataset}{Data frame or matrix to use to find row weights.}

\item{target}{Numeric vector of target column means.}

\item{max_iter}{Maximum number of iterations for \code{optim()}.}
}
\description{
Compute edge weights \code{p}
such that the column means of
\code{tboot(dataset = dataset, p = p)}
equal \code{target} on average.
}
\details{
This function minimizes the Kullback-Leibler divergence
between two sampling schemes:
 (P_C) classic resampling with each observation sampled with equal prob
 (P_W) weighted resampling with each sample i getting a different prob (p_i)
      D(P_C||P_W)=sum_i log(1/n) - log(p_i)
The function finds p_i such that:
      sum_i p_i = 1
      dataset' p = target
  where dataset is a N x K matrix of clinical variables.
  Optimization is perfromed by utilizing lagrange multipliers
  Output is the optimal porbability (p)
}
\seealso{
\code{\link{tboot}}
}
