% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweights.R
\name{tweights}
\alias{tweights}
\title{Function \code{tweights}}
\usage{
tweights(dataset, target = apply(dataset, 2, mean), distance = "euchlidean",
  control = list(maxit = 10000, reltol = 1e-16), tol = 1e-05,
  silent = FALSE)
}
\arguments{
\item{dataset}{Data frame or matrix to use to find row weights.}

\item{target}{Numeric vector of target column means.}

\item{distance}{The distance to minimize. Must be either 'euchlidean' or 'kl' (i.e. Kullback-Leibler).}

\item{control}{Controll prameter to be passed into \code{optConstr()}.}

\item{tol}{Tolerance. If the achieved mean is to far from the target (i.e. as defined by tol) an error will be thrown.}

\item{silent}{Allows silencing some messages.}
}
\description{
Returns a vector \code{p} of resampling probabilities 
such that the column means of \code{tboot(dataset = dataset, p = p)}
equals \code{target} on average.
}
\details{
Let \eqn{p_i = 1/n} be  probability of sampling subject \eqn{i} from a dataset with \eqn{n} individuals (i.e. rows of the dataset) in the classic resampling with replacement scheme.
Also, let \eqn{q_i} be the probability of sampling subject \eqn{i} from a dataset with \eqn{n} individuals in our new resampling scheme. Let \eqn{d(q,p)} represent a distance between the two resampling schemes.  The \code{tweights}
function seeks to solve the problem: 
\deqn{q = argmin_p d(q,p)}
Subject to the constraint that:
\deqn{ sum_i q_i = 1} and
\deqn{  dataset' q = target}
where dataset is a n x K matrix of variables input to the function.

  \deqn{d_euclidian(q,p) = sqrt( sum_i (p_i-q_i)^2 )}
  \deqn{d_kl(q,p) = sum_i (log(p_i) - log(q_i))}

Optimization for euclidean distance is a quadratic program and utilizes the ipop function in kernLab.
The euclidean based solution helps form a starting value which is used along with the constOptim function 
and lagrange multipliers to solve the Kullback-Leibler distance optimization.
  Output is the optimal porbability (p)
}
\seealso{
\code{\link{tboot}}
}
