% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweights_bmr.R
\name{tweights_bmr}
\alias{tweights_bmr}
\title{Function tweights_bmr}
\usage{
tweights_bmr(dataset, marginal, distance = "klqp", maxit = 1000,
  tol = 1e-08, warningcut = 0.05, silent = FALSE, Nindependent = 1,
  augmentWeights = NULL)
}
\arguments{
\item{dataset}{Data frame or matrix to use to find row weights.}

\item{marginal}{Must be a named list with each element a vector of simulations of the marginal distribution of the posterior mean of data in the dataset.}

\item{distance}{The distance measure to minimize. Must be either 'euchlidean' or 'kl' (i.e. Kullback-Leibler). 'klqp' is recomneded.}

\item{maxit}{Defines the maximum number of iterations for optimizing 'kl' distance.}

\item{tol}{Tolerance. If the achieved mean is to far from the target (i.e. as defined by tol) an error will be thrown.}

\item{warningcut}{Sets the cutoff for determining when a large weight will trigger a warnint.}

\item{silent}{Allows silencing some messages.}

\item{Nindependent}{Assumes the input also includes 'Nindependent'samples with independent columns. See details.}

\item{augmentWeights}{List with weights for each variable marginal. Only has effect if Nindependent!=0. Generally this should be left as NULL. 'tweights' will fill in automatically.}
}
\description{
Setup the needed prequisitesin order to prepare for bayesian marginal reconstruction (including a call to tweights). Takes as input simulations from the posterior marginal distribution of variables in a dataset.
}
\seealso{
\code{\link{tweights}}
}
